avoid unnecessary _sc_ensure calls by late patch
During frame lowering the stack size of callees is not known, it thus takes a
module pass to do an easy optimization and avoid ensures after calls to
functions that do not reserve a stack (eg. most leaves).

PatmosStackCacheAnalysis is already set up to run at the right time, so
emitting ensures can be simply done there. The pass is now enabled by default (ie. hack).