singlepath: Fixed an issue with how predicate definition instruction were emitted.
When two blocks are bundled, their predicate defintions are added to the resulting block.
The two original blocks might be defining the same abstract predicate, meaning the resulting
block had two definitions for the same predicate. In such a case, both definitions would
be emitted using unguarded PAND instructions. This is wrong, as the second one will then
always overwrite the first, even though it shouldn't.

This fix solves the problem by ensuring any defintion for a predicate after the first
will use a guarded PMOV instruction instead.

This was not previously a problem for non-bundled blocks because the use of guarded PMOV
was already used for the second block defining a predicate. So this is the same solution
but for bundled blocks.