Made inline assembly parsing less intrusive.
The parsing of prefixes ('{' for bundles and '(..)' for guards)
was previously implemented by adding custom support for it in the
llvm tablegen and parser code.

This commit changes the implementation to be less dependent on changes
to llvm files.
The TableGen backends 'AsmMatcherEmitter' and 'AsmWriterEmitter' have
been reverted to their original states for this version of llvm (3.4.2).
As has 'include/llvm/Target/Target.td'.
'lib/MC/MCParser/AsmParser.ccp' has also been reverted,
with the exception of adding '{' and '(' as allowed identifiers.
The first of these is already allowed in llvm verson 11.0.0,
while the second is not.

This new implementation still requires custom support in 'AsmParser.cpp'
(for '('), however, this is much less intrusive than previously.