Added testing of -mpatmos-max-subfunction-size flag.
Moved the '%test_no_runtime_execution' llvm-lit command
to the root Patmos test folder to make it usable by all other
folders.

Added the '%XFAIL-filecheck' llvm-lit command that makes it
easy to test failures and their messages at the same time.
This is a more robust way to test, as checking the error
message ensures the correct error is caught.
With XFAIL, if something changes that invalidates a test
but happens to also issue a failure, this would not be caught
as the test is reported as a success (XFAIL).
With %XFAIL-filecheck, if the error changes into something
unexpected, the test is reported as a failure.

Edited previous XFAIL tests to use %XFAIL-filecheck